voice-backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
openai-whisper
pyannote.audio==3.3.0
torch==2.2.0
torchaudio==2.2.0
torchvision
ffmpeg-python
noisereduce
soundfile
numpy